Fix SupportSkillParser, SkillsParser and ItemParser results not being cached
